a86/interpreter/state: rewrite state initialization
The `initialize-state` function now supports an `#:entry-point` keyword argument
for specifying the label at which program execution should begin. This is then
used to generate additional instructions as a wrapper, to simulate the provided
code being called from somewhere else.

A new `sequence` function was added to `a86/interpreter/utility` to accommodate
the instruction generation. This is identical to the `seq` function from the 430
course's a86 language.