a86/interpreter/interp: revise argument handling and implement Global
The `process-argument` inner function to `step` will now specify what kind of
instruction is to blame in the event of an error. It has also been extended to
support an `'external-function` argument processing mechanism, so that all
arguments can be processed with this function.

The `Global` instruction implementation mirrors that of `Label`, i.e., it does
nothing, simply proceeding to the next instruction.

There was an error in the implementation of `Call` which was fetching the values
from memory at the address that ought to be returned to and storing the result
in the instruction pointer, leading to invalid execution. This has been correct
by using `process-argument` appropriately.