a86/*: overhaul interpreter implementation
This commit replaces the previous memory implementation with a much more fully
featured version. We now have proper emulation for the various sections (.text,
.rodata, .data, .bss) as well as the stack and heap. Read/write permissions are
properly enforced within these sections as would be expected.

The implementation of the dynamically sized regions of memory (the stack and
heap) are lightly optimized. Although their address ranges are fairly large,
they are implemented as vectors that automatically resize when writes are
attempted beyond their maximum indices.

The instructions that are passed to the interpreter are automatically separated
into their appropriate sections, with new options for the .rodata and .bss
sections that do not exist in upstream a86.

Light testing suggests the bulk of the implementation works as expected, but
further testing is required and recommended.