Rewrite memory-ref/memory-set! to properly handle unaligned accesses
The memory reading/writing processes can now be given non-word-aligned addresses
and are expected to work correctly. The [memory-ref] function will always
produce a word as a result, but that word can be read from across two other
words. The [memory-set!] function can now write less than a word's worth of
bytes, and the write can happen across a word boundary. It now takes an optional
[byte-count] argument to specify how far to write.