Implement symbol-manipulation functions for Mug runtime
This is part of #16, which is actually about Mug rather than Mountebank. Oops.

Also includes new [runtime/memory-ref] and [runtime/memory-set!] functions that
are made available within the bodies of runtime functions. This avoids the
client having to think about time ticks or getting access to the emulator's
memory directly.