Replace previous emulation loop
The previous emulation loop system was incredibly complex and brittle, and it
also did not permit extensibility in the desired way. The new system solves
these problems to some degree, though it is not perfect.

An evaluation loop is extended by the [define-loop-command] directive (from
[a86/emulate/command-loop.rkt]). A loop command is a (set of) function(s) that
performs some side-effects and then indirectly calls another loop command. This
should allow dynamic extension for use by the REPL system.